+++ ui/src/components/records/RecordsPicker.svelte
@@ -214,7 +214,7 @@
         {#if !isView}
             <button
                 type="button"
-                class="btn btn-transparent btn-hint p-l-sm p-r-sm"
+                class="btn btn-pill btn-transparent btn-hint p-l-xs p-r-xs"
                 on:click={() => upsertPanel?.show()}
             >
                 <div class="txt">New record</div>